boost removed

+++ yasmin_ros/include/yasmin_ros/monitor_state.hpp
@@ -111,7 +111,7 @@ private:
     if (this->monitoring) {
       this->msg_list.push_back(*msg.get());
 
-      if (this->msg_list.size() >= this->msg_queue) {
+      if ((int)this->msg_list.size() >= this->msg_queue) {
         this->msg_list.erase(this->msg_list.begin());
       }
     }
